refactor(typescript): organize typescript files and types
- Refactored TypeScript files and types into a dedicated directory for better organization.
- Converted types into declaration files to avoid the need for explicit imports in components.
- Added a TypeScript check script to 'package.json' for linting and type checking.
- Configured pre-commit hook to run TypeScript checks before committing code.
- Updated README.md to reflect the changes in the TypeScript setup.
