+++ package.json
@@ -7,7 +7,10 @@
     "build": "next build",
     "start": "next start",
     "lint": "next lint",
-    "prepare": "husky"
+    "prepare": "husky",
+    "check-types": "tsc --noEmit",
+    "precommit": "run-s check-types process-staged-files",
+    "process-staged-files": "lint-staged"
   },
   "dependencies": {
     "@heroicons/react": "^2.1.5",
@@ -33,6 +36,7 @@
     "eslint-plugin-tailwindcss": "^3.17.4",
     "husky": "^9.1.6",
     "lint-staged": "^15.2.10",
+    "npm-run-all": "^4.1.5",
     "postcss": "^8",
     "prettier": "^3.3.3",
     "tailwindcss": "^3.4.1",
