+++ src/config/constants.ts
@@ -4,6 +4,8 @@ export const PAGE_SIZE = 10;
 
 export const DEFAULT_PAGE = 1;
 
+export const CONTENT_TYPE_APPLICATION_JSON = "application/json";
+
 export const ROUTES = [
   {
     path: "/",
