+++ src/lib/api-clients/newYorkTimesApiClient.ts
@@ -1,10 +1,9 @@
 import axios from "axios";
 
-import { ApiError } from "@/types/api";
+import { CONTENT_TYPE_APPLICATION_JSON } from "@/config/constants";
 
 const NEW_YORK_TIMES_API_KEY = process.env.NEW_YORK_TIMES_API_KEY;
 const NEW_YORK_TIMES_API_BASE_URL = process.env.NEW_YORK_TIMES_API_BASE_URL;
-const CONTENT_TYPE_APPLICATION_JSON = "application/json";
 
 const newYorkTimesApiClient = axios.create({
   baseURL: NEW_YORK_TIMES_API_BASE_URL,
