+++ src/lib/api-clients/newsApiClient.ts
@@ -1,10 +1,9 @@
 import axios from "axios";
 
-import { ApiError } from "@/types/api";
+import { CONTENT_TYPE_APPLICATION_JSON } from "@/config/constants";
 
 const NEWS_API_KEY = process.env.NEWS_API_KEY;
 const NEWS_API_BASE_URL = process.env.NEWS_API_BASE_URL;
-const CONTENT_TYPE_APPLICATION_JSON = "application/json";
 
 const newsApiClient = axios.create({
   baseURL: NEWS_API_BASE_URL,
